team tab added

+++ routes/deletespeaker.js
@@ -9,12 +9,16 @@ const config = require('../config');
 const con = mysql.createConnection(config.MYSQL);
 
 router.post('/', function(req, res, next)  {
-    const post = req.body;
-    const speakerName  = sanitizeHtml(post.speakerName);
-    var sql = "Delete from speaker where name = '" + speakerName +"'" ;
-    con.query(sql, function (err, result, fields) {
-        res.redirect('/admin');
-    });
+    if(!req.session.username) {
+        res.redirect("http://tedxnitkurukshetra.com");
+    }else {
+        const post = req.body;
+        const speakerName = sanitizeHtml(post.speakerName);
+        var sql = "Delete from speaker where name = '" + speakerName + "'";
+        con.query(sql, function (err, result, fields) {
+            res.redirect('/admin');
+        });
+    }
 });
 
 module.exports = router;
